+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod command;
 pub mod context;
 pub mod database;
+pub mod handler;
 pub mod models;
