+++ src/main.rs
@@ -1,100 +1,8 @@
-use rexmit::{database::{
-    clear_guild_queue, get_guild_is_subscribed, pop_guild_queue, set_guild_queue,
-    set_joined_to_channel, count_subscribed_guilds_joined_to_channel, count_guilds_joined_to_channel, count_free_guilds_joined_to_channel,
-}, command::{check_msg, unmute, UNMUTE_COMMAND}};
-use serenity::{
-    async_trait,
-    client::{Cache, Client, Context, EventHandler},
-    framework::{
-        standard::{
-            macros::{command, group},
-            Args, CommandResult,
-        },
-        StandardFramework,
-    },
-    http::Http,
-    model::{
-        channel::Message,
-        gateway::Ready,
-        prelude::{Activity, ChannelId},
-    },
-    prelude::{GatewayIntents, Mentionable},
-    Result as SerenityResult,
-};
-use songbird::{
-    input::{self, restartable::Restartable},
-    Event, EventContext, EventHandler as VoiceEventHandler, SerenityInit, TrackEvent,
-};
-use std::{env, sync::Arc, time::Duration};
-use tracing::{Level, info};
-
-struct Handler;
-
-#[async_trait]
-impl EventHandler for Handler {
-    async fn ready(&self, ctx: Context, ready: Ready) {
-        ctx.set_activity(Activity::listening("~q <youtube url>"))
-            .await;
-        println!("{} is connected!", ready.user.name);
-
-        let free_guilds_option = count_free_guilds_joined_to_channel().await;
-        if free_guilds_option.is_some() {
-            info!("free guilds{}", free_guilds_option.unwrap())
-        }
-        
-        let sub_guilds_option = count_subscribed_guilds_joined_to_channel().await;
-        if sub_guilds_option.is_some() {
-            info!("sub guilds: {}", sub_guilds_option.unwrap())
-        }
-
-        let guilds_option = count_guilds_joined_to_channel().await;
-        if guilds_option.is_some() {
-            info!("total guilds: {}", guilds_option.unwrap())
-        }
-        
-        //checking guild queues, could use better naming
-        /*let joined_guilds_option = get_guilds_joined_to_channel().await;
-        if joined_guilds_option.is_some() {
-            let joined_guilds = joined_guilds_option.unwrap();
-            for joined_guild in joined_guilds {
-                let subscribed_option = get_guild_is_subscribed(joined_guild.to_string()).await;
-                match subscribed_option {
-                    Some(subscribed) => {
-                        info!("guild is subscribed: {:?}", subscribed);
-                    },
-                    None => {}
-                }
-            }
-        }*/
-    }
-
-    /*async fn voice_state_update(&self, _ctx: Context, _old: Option<VoiceState>, _new: VoiceState) {
-        let channel_result = _ctx.http.get_channel(_new.channel_id.unwrap().into()).await;
-        match channel_result.unwrap().guild() {
-            Some(guild_channel) => {
-                let members_result = guild_channel.members(&_ctx.cache).await;
-                match members_result {
-                    Ok(members) => {
-                        println!("{}", members.len())
-                    },
-                    Err(why) => {
-                        println!("{}", why);
-                    }
-                }
-            },
-            None => {
-                println!("{}", "no guild")
-            }
-        };
-    }*/
-}
-
-#[group]
-#[commands(
-    d, deafen, j, join, l, leave, m, mute, q, queue, s, skip, c, clear, stop, p, ping, ud,
-    undeafen, um, unmute
-)]
-struct General;
+use rexmit::{command::GENERAL_GROUP, handler::Handler};
+use serenity::{client::Client, framework::StandardFramework, prelude::GatewayIntents};
+use songbird::SerenityInit;
+use std::env;
+use tracing::Level;
 
 #[tokio::main]
 async fn main() {
@@ -131,719 +39,4 @@ async fn main() {
         .start()
         .await
         .map_err(|why| println!("Client ended: {:?}", why));
-}
-
-#[command]
-async fn d(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return deafen(ctx, msg, _args).await;
-}
-
-#[command]
-async fn deafen(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    let handler_lock = match manager.get(guild_id) {
-        Some(handler) => handler,
-        None => {
-            check_msg(msg.reply(ctx, "Not in a voice channel").await);
-
-            return Ok(());
-        }
-    };
-
-    let mut handler = handler_lock.lock().await;
-
-    if handler.is_deaf() {
-        check_msg(msg.channel_id.say(&ctx.http, "Already deafened").await);
-    } else {
-        if let Err(e) = handler.deafen(true).await {
-            check_msg(
-                msg.channel_id
-                    .say(&ctx.http, format!("Failed: {:?}", e))
-                    .await,
-            );
-        }
-
-        check_msg(msg.channel_id.say(&ctx.http, "Deafened").await);
-    }
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn j(ctx: &Context, msg: &Message) -> CommandResult {
-    return join(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn join(ctx: &Context, msg: &Message) -> CommandResult {
-    if msg.guild_id.is_some() {
-        let og_guild = msg.guild(&ctx.cache).unwrap();
-        let guild = og_guild.clone();
-        let guild_id = guild.id;
-        let subscribed_option = get_guild_is_subscribed(guild_id.to_string()).await;
-        match subscribed_option {
-            Some(subscribed) => {
-                if subscribed {
-                    let channel_id = guild
-                        .voice_states
-                        .get(&msg.author.id)
-                        .and_then(|voice_state| voice_state.channel_id);
-
-                    let connect_to = match channel_id {
-                        Some(channel) => channel,
-                        None => {
-                            check_msg(msg.reply(ctx, "Not in a voice channel").await);
-
-                            return Ok(());
-                        }
-                    };
-
-                    let manager = songbird::get(ctx)
-                        .await
-                        .expect("Songbird Voice client placed in at initialisation.")
-                        .clone();
-
-                    let (handle_lock, success) = manager.join(guild_id, connect_to).await;
-
-                    if let Ok(_channel) = success {
-                        check_msg(
-                            msg.channel_id
-                                .say(&ctx.http, &format!("Joined {}", connect_to.mention()))
-                                .await,
-                        );
-
-                        let guild_id = msg.guild_id.unwrap();
-
-                        let chan_id = msg.channel_id;
-
-                        let send_http = ctx.http.clone();
-
-                        let mut handle = handle_lock.lock().await;
-
-                        handle.add_global_event(
-                            Event::Track(TrackEvent::End),
-                            TrackEndNotifier {
-                                guild,
-                                chan_id,
-                                http: send_http,
-                            },
-                        );
-
-                        let send_http = ctx.http.clone();
-                        let send_cache = ctx.cache.clone();
-
-                        handle.add_global_event(
-                            Event::Periodic(Duration::from_secs(1800), None),
-                            Periodic {
-                                voice_channel_id: connect_to,
-                                message_channel_id: chan_id,
-                                http: send_http,
-                                cache: send_cache,
-                                ctx: ctx.clone(),
-                            },
-                        );
-
-                        set_joined_to_channel(
-                            guild_id.to_string(),
-                            Some(connect_to.to_string()),
-                            Some(chan_id.to_string()),
-                        )
-                        .await;
-                    } else {
-                        check_msg(
-                            msg.channel_id
-                                .say(&ctx.http, "Error joining the channel")
-                                .await,
-                        );
-                    }
-                } else {
-                    check_msg(
-                        msg.channel_id
-                            .say(&ctx.http, "oops! no subscription active")
-                            .await,
-                    );
-                }
-            }
-            None => {
-                check_msg(msg.channel_id.say(&ctx.http, "oops! no db!").await);
-            }
-        }
-    }
-    Ok(())
-}
-
-struct TrackEndNotifier {
-    guild: serenity::model::prelude::Guild,
-    chan_id: ChannelId,
-    http: Arc<Http>,
-}
-
-#[async_trait]
-impl VoiceEventHandler for TrackEndNotifier {
-    async fn act(&self, ctx: &EventContext<'_>) -> Option<Event> {
-        if let EventContext::Track(track_list) = ctx {
-            pop_guild_queue(self.guild.id.to_string()).await;
-            check_msg(
-                self.chan_id
-                    .say(
-                        &self.http,
-                        &format!(
-                            "Track ended: {}",
-                            track_list
-                                .first()
-                                .as_ref()
-                                .unwrap()
-                                .1
-                                .metadata()
-                                .source_url
-                                .as_ref()
-                                .unwrap()
-                        ),
-                    )
-                    .await,
-            );
-        }
-
-        None
-    }
-}
-
-#[command]
-#[only_in(guilds)]
-async fn l(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return leave(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn leave(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    if msg.guild_id.is_some() {
-        let guild = msg.guild(&ctx.cache).unwrap();
-        let guild_id = guild.id;
-
-        let manager = songbird::get(ctx)
-            .await
-            .expect("Songbird Voice client placed in at initialisation.")
-            .clone();
-        let has_handler = manager.get(guild_id).is_some();
-
-        if has_handler {
-            if let Err(e) = manager.remove(guild_id).await {
-                check_msg(
-                    msg.channel_id
-                        .say(&ctx.http, format!("Failed: {:?}", e))
-                        .await,
-                );
-            }
-
-            check_msg(msg.channel_id.say(&ctx.http, "Left voice channel").await);
-            clear_guild_queue(guild_id.to_string()).await;
-            set_joined_to_channel(guild_id.to_string(), None, None).await;
-        } else {
-            check_msg(msg.reply(ctx, "Not in a voice channel").await);
-        }
-    }
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn m(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return mute(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn mute(ctx: &Context, msg: &Message) -> CommandResult {
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    let handler_lock = match manager.get(guild_id) {
-        Some(handler) => handler,
-        None => {
-            check_msg(msg.reply(ctx, "Not in a voice channel").await);
-
-            return Ok(());
-        }
-    };
-
-    let mut handler = handler_lock.lock().await;
-
-    if handler.is_mute() {
-        check_msg(msg.channel_id.say(&ctx.http, "Already muted").await);
-    } else {
-        if let Err(e) = handler.mute(true).await {
-            check_msg(
-                msg.channel_id
-                    .say(&ctx.http, format!("Failed: {:?}", e))
-                    .await,
-            );
-        }
-
-        check_msg(msg.channel_id.say(&ctx.http, "Now muted").await);
-    }
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn p(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return ping(ctx, msg, _args).await;
-}
-
-#[command]
-async fn ping(ctx: &Context, msg: &Message) -> CommandResult {
-    check_msg(msg.channel_id.say(&ctx.http, "Pong!").await);
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn play_fade(ctx: &Context, msg: &Message, mut args: Args) -> CommandResult {
-    let url = match args.single::<String>() {
-        Ok(url) => url,
-        Err(_) => {
-            check_msg(
-                msg.channel_id
-                    .say(&ctx.http, "Must provide a URL to a video or audio")
-                    .await,
-            );
-
-            return Ok(());
-        }
-    };
-
-    if !url.starts_with("http") {
-        check_msg(
-            msg.channel_id
-                .say(&ctx.http, "Must provide a valid URL")
-                .await,
-        );
-
-        return Ok(());
-    }
-
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    if let Some(handler_lock) = manager.get(guild_id) {
-        let mut handler = handler_lock.lock().await;
-
-        let source = match input::ytdl(&url).await {
-            Ok(source) => source,
-            Err(why) => {
-                println!("Err starting source: {:?}", why);
-
-                check_msg(msg.channel_id.say(&ctx.http, "Error sourcing ffmpeg").await);
-
-                return Ok(());
-            }
-        };
-
-        // This handler object will allow you to, as needed,
-        // control the audio track via events and further commands.
-        let song = handler.play_source(source);
-        let send_http = ctx.http.clone();
-        let chan_id = msg.channel_id;
-
-        // This shows how to periodically fire an event, in this case to
-        // periodically make a track quieter until it can be no longer heard.
-        let _ = song.add_event(
-            Event::Periodic(Duration::from_secs(5), Some(Duration::from_secs(7))),
-            SongFader {
-                chan_id,
-                http: send_http,
-            },
-        );
-
-        let send_http = ctx.http.clone();
-
-        // This shows how to fire an event once an audio track completes,
-        // either due to hitting the end of the bytestream or stopped by user code.
-        let _ = song.add_event(
-            Event::Track(TrackEvent::End),
-            SongEndNotifier {
-                chan_id,
-                http: send_http,
-            },
-        );
-
-        check_msg(msg.channel_id.say(&ctx.http, "Playing song").await);
-    } else {
-        check_msg(
-            msg.channel_id
-                .say(&ctx.http, "Not in a voice channel to play in")
-                .await,
-        );
-    }
-
-    Ok(())
-}
-
-struct SongFader {
-    chan_id: ChannelId,
-    http: Arc<Http>,
-}
-
-#[async_trait]
-impl VoiceEventHandler for SongFader {
-    async fn act(&self, ctx: &EventContext<'_>) -> Option<Event> {
-        if let EventContext::Track(&[(state, track)]) = ctx {
-            let _ = track.set_volume(state.volume / 2.0);
-
-            if state.volume < 1e-2 {
-                let _ = track.stop();
-                check_msg(self.chan_id.say(&self.http, "Stopping song...").await);
-                Some(Event::Cancel)
-            } else {
-                check_msg(self.chan_id.say(&self.http, "Volume reduced.").await);
-                None
-            }
-        } else {
-            None
-        }
-    }
-}
-
-struct SongEndNotifier {
-    chan_id: ChannelId,
-    http: Arc<Http>,
-}
-
-#[async_trait]
-impl VoiceEventHandler for SongEndNotifier {
-    async fn act(&self, _ctx: &EventContext<'_>) -> Option<Event> {
-        check_msg(
-            self.chan_id
-                .say(&self.http, "Song faded out completely!")
-                .await,
-        );
-
-        None
-    }
-}
-
-struct Periodic {
-    voice_channel_id: ChannelId,
-    message_channel_id: ChannelId,
-    http: Arc<Http>,
-    cache: Arc<Cache>,
-    ctx: Context,
-}
-
-#[async_trait]
-impl VoiceEventHandler for Periodic {
-    async fn act(&self, _ctx: &EventContext<'_>) -> Option<Event> {
-        let channel = self.http.get_channel(self.voice_channel_id.into()).await;
-        match channel.unwrap().guild() {
-            Some(guild_channel) => {
-                let members = guild_channel.members(&self.cache).await;
-
-                // please modularize this monstrocity
-                // what i mean by this is create some functions and call the functions instead
-                // we want to utilize DRY (DON'T REPEAT YOURSELF) principles
-                if members.unwrap().len() <= 1 {
-                    let manager = songbird::get(&self.ctx)
-                        .await
-                        .expect("Songbird Voice client placed in at initialisation.")
-                        .clone();
-
-                    let has_handler = manager.get(guild_channel.guild_id).is_some();
-
-                    if has_handler {
-                        if let Err(e) = manager.remove(guild_channel.guild_id).await {
-                            check_msg(
-                                self.message_channel_id
-                                    .say(&self.http, format!("Failed: {:?}", e))
-                                    .await,
-                            );
-                        }
-
-                        check_msg(
-                            self.message_channel_id
-                                .say(&self.http, "Left voice channel")
-                                .await,
-                        );
-                        match guild_channel.guild(&self.cache) {
-                            Some(_guild) => {
-                                clear_guild_queue(guild_channel.guild_id.to_string()).await;
-                            }
-                            None => {}
-                        };
-                        set_joined_to_channel(guild_channel.guild_id.to_string(), None, None).await;
-                    } else {
-                        check_msg(
-                            self.message_channel_id
-                                .say(&self.http, "Not in a voice channel")
-                                .await,
-                        );
-                    }
-                }
-            }
-            None => {
-                println!("{}", "channel is none")
-            }
-        }
-
-        None
-    }
-}
-
-#[command]
-#[only_in(guilds)]
-async fn q(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
-    return queue(ctx, msg, args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn queue(ctx: &Context, msg: &Message, mut args: Args) -> CommandResult {
-    // capture args before mutating
-    let og_args = &args.clone();
-
-    let url = match args.single::<String>() {
-        Ok(url) => url,
-        Err(_) => {
-            check_msg(
-                msg.channel_id
-                    .say(&ctx.http, "Must provide a URL to a video or audio")
-                    .await,
-            );
-
-            return Ok(());
-        }
-    };
-
-    if !url.starts_with("http") {
-        check_msg(
-            msg.channel_id
-                .say(&ctx.http, "Must provide a valid URL")
-                .await,
-        );
-
-        return Ok(());
-    }
-
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    if let Some(handler_lock) = manager.get(guild_id) {
-        let mut handler = handler_lock.lock().await;
-
-        // Here, we use lazy restartable sources to make sure that we don't pay
-        // for decoding, playback on tracks which aren't actually live yet.
-        let source = match Restartable::ytdl(url, true).await {
-            Ok(source) => source,
-            Err(why) => {
-                println!("Err starting source: {:?}", why);
-
-                check_msg(msg.channel_id.say(&ctx.http, "Error sourcing ffmpeg").await);
-
-                return Ok(());
-            }
-        };
-
-        handler.enqueue_source(source.into());
-
-        check_msg(
-            msg.channel_id
-                .say(
-                    &ctx.http,
-                    format!("Added song to queue: position {}", handler.queue().len()),
-                )
-                .await,
-        );
-
-        let mut queue = vec![];
-        for track_handle in handler.queue().current_queue() {
-            queue.push(track_handle.metadata().source_url.clone().unwrap())
-        }
-
-        set_guild_queue(guild_id.to_string(), queue).await;
-    } else {
-        check_msg(
-            msg.channel_id
-                .say(
-                    &ctx.http,
-                    "not in a voice channel to play in. attempting to join",
-                )
-                .await,
-        );
-        match join(ctx, msg, og_args.to_owned()).await {
-            Ok(result) => {
-                match queue(ctx, msg, og_args.to_owned()).await {
-                    Ok(result) => result,
-                    Err(_why) => {
-                        check_msg(msg.channel_id.say(&ctx.http, "unable to queue").await);
-                    }
-                }
-                return Ok(result);
-            }
-            Err(_why) => {
-                check_msg(msg.channel_id.say(&ctx.http, "unable to join").await);
-                return Ok(());
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn s(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return skip(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn skip(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    if let Some(handler_lock) = manager.get(guild_id) {
-        let handler = handler_lock.lock().await;
-        let track_queue = handler.queue();
-        let _ = track_queue.skip();
-
-        check_msg(
-            msg.channel_id
-                .say(
-                    &ctx.http,
-                    format!("Song skipped: {} in queue.", track_queue.len()),
-                )
-                .await,
-        );
-    } else {
-        check_msg(
-            msg.channel_id
-                .say(&ctx.http, "Not in a voice channel to play in")
-                .await,
-        );
-    }
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn c(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return stop(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn clear(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return stop(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn stop(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    if let Some(handler_lock) = manager.get(guild_id) {
-        let handler = handler_lock.lock().await;
-        let track_queue = handler.queue();
-        let _ = track_queue.stop();
-
-        check_msg(msg.channel_id.say(&ctx.http, "Queue cleared.").await);
-
-        let mut queue = vec![];
-        for track_handle in handler.queue().current_queue() {
-            queue.push(track_handle.metadata().source_url.clone().unwrap())
-        }
-
-        clear_guild_queue(guild_id.to_string()).await;
-    } else {
-        check_msg(
-            msg.channel_id
-                .say(&ctx.http, "Not in a voice channel to play in")
-                .await,
-        );
-    }
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn ud(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return undeafen(ctx, msg, _args).await;
-}
-
-#[command]
-#[only_in(guilds)]
-async fn undeafen(ctx: &Context, msg: &Message) -> CommandResult {
-    let guild = msg.guild(&ctx.cache).unwrap();
-    let guild_id = guild.id;
-
-    let manager = songbird::get(ctx)
-        .await
-        .expect("Songbird Voice client placed in at initialisation.")
-        .clone();
-
-    if let Some(handler_lock) = manager.get(guild_id) {
-        let mut handler = handler_lock.lock().await;
-        if let Err(e) = handler.deafen(false).await {
-            check_msg(
-                msg.channel_id
-                    .say(&ctx.http, format!("Failed: {:?}", e))
-                    .await,
-            );
-        }
-
-        check_msg(msg.channel_id.say(&ctx.http, "Undeafened").await);
-    } else {
-        check_msg(
-            msg.channel_id
-                .say(&ctx.http, "Not in a voice channel to undeafen in")
-                .await,
-        );
-    }
-
-    Ok(())
-}
-
-#[command]
-#[only_in(guilds)]
-async fn um(ctx: &Context, msg: &Message, _args: Args) -> CommandResult {
-    return unmute(ctx, msg, _args).await;
 }
